[AArch64][FEAT_CMPBR] Codegen for Armv9.6-a CBB and CBH
This patch adds codegen for CBB and CBH, CB variants operating on bytes
and half-words instead of full registers, allowing us to fold sign- and
zero extensions.

Since a number of optimizations prefer explicit branches, we need to be
able to unfold back to a sequence of extensions and cmp. To achieve
this, the CBB and CBH pseudos carry the exention type of their register
operands as last operands.